Update dev dependencies to appease security warnings.
Urllib3 had a security vulnerability, which was flagged by GitHub's security
alerts. As this was a dependency of twine, a dev dependency, this did not
really affect the library.

Unfortunately, because of https://github.com/pypa/pipenv/issues/2412 there is
no functional "--selective-upgrade" option for pipenv and I could not just
change the lockfile for that one thing, and so the other dependencies have
been updated as well. It will be a little bit out-of-sync with what has been
released (in that the latest was built against pyyaml 5.1.1 and not 5.1), but
this is okay.